requirements.txt
WeasyPrint
psycopg2-binary 
django-bootstrap-datepicker-plus
django-crispy-forms
django